requirements.txt
flask>=2.3.0
gunicorn>=21.0.0
flask-pymongo>=2.3.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
matplotlib>=3.7.0
shapely>=2.0.0
networkx>=3.1.0
pytz>=2023.3
python-dotenv>=1.0.0
pymongo>=4.5.0
click>=8.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
# Security dependencies
pyjwt>=2.8.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
cryptography>=41.0.0
bcrypt>=4.0.0